+++ Assets/Samples/SimpleDemo/SimpleControls.cs
@@ -99,7 +99,8 @@ public partial class @SimpleControls: IInputActionCollection2, IDisposable
                     ""expectedControlType"": ""Button"",
                     ""processors"": """",
                     ""interactions"": ""Tap,SlowTap"",
-                    ""initialStateCheck"": false
+                    ""initialStateCheck"": false,
+                    ""priority"": 0
                 },
                 {
                     ""name"": ""move"",
@@ -108,7 +109,8 @@ public partial class @SimpleControls: IInputActionCollection2, IDisposable
                     ""expectedControlType"": ""Vector2"",
                     ""processors"": """",
                     ""interactions"": """",
-                    ""initialStateCheck"": true
+                    ""initialStateCheck"": true,
+                    ""priority"": 0
                 },
                 {
                     ""name"": ""look"",
@@ -117,7 +119,8 @@ public partial class @SimpleControls: IInputActionCollection2, IDisposable
                     ""expectedControlType"": ""Vector2"",
                     ""processors"": """",
                     ""interactions"": """",
-                    ""initialStateCheck"": true
+                    ""initialStateCheck"": true,
+                    ""priority"": 0
                 }
             ],
             ""bindings"": [